Convert to unsigned int before left shift
This is to fix overflow when 128 is left shifted by 24.

+++ vpx_ports/mem_ops.h
@@ -89,10 +89,10 @@ static unsigned MEM_VALUE_T mem_get_be32(const void *vmem) {
   unsigned MEM_VALUE_T  val;
   const MAU_T          *mem = (const MAU_T *)vmem;
 
-  val = mem[0] << 24;
-  val |= mem[1] << 16;
-  val |= mem[2] << 8;
-  val |= mem[3];
+  val = ((unsigned int)mem[0]) << 24;
+  val |= ((unsigned int)mem[1]) << 16;
+  val |= ((unsigned int)mem[2]) << 8;
+  val |= ((unsigned int)mem[3]);
   return val;
 }
